:: CRAFTY_CoBRA_Trunk v0.2.33 :: - set to version 0.2.33 - feature added: probabilistic choice for giving up (DefaultLandUseAgent#considerGivingUp(), added attribute givingUpProbabilityWeight to AbstractFR, added Region#getMaxGivingUpThresholdDeviation(), added AgentProperyIds#GIVING_UP_PROB_WEIGHT - medium change: changed type from integer to float for GiveUpGiveInAllocationModel.probabilityExponent - medium change: check for incompatible Demand model in Region#preTick() - medium change: adaptation in IntegratedTest - minor change: comments in AdaptionCurveCompetitivenessModel and GlobalCompFuncSubsidyPa

+++ src/org/volante/abm/example/GiveUpGiveInAllocationModel.java
@@ -118,7 +118,7 @@ public class GiveUpGiveInAllocationModel extends SimpleAllocationModel
 	 * to 0 for uniformly random selection.
 	 */
 	@Attribute(required = false)
-	public int				probabilityExponent	= 2;
+	public float probabilityExponent = 2.0f;
 
 	protected Cell						perfectCell			= new Cell();
 	protected ModelData					data				= null;
